pyproject.toml
[project]
name = "gworkspace-mcp"
version = "0.1.3"
description = "Google Workspace MCP Server - Connect Claude to Google Workspace APIs"
authors = [
    {name = "Masa", email = "github@matsuoka.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["mcp", "google-workspace", "gmail", "calendar", "drive", "docs", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.1",
    "google-api-python-client>=2.158.0",
    "httpx>=0.28.1",
    "mcp>=1.3.1",
    "click>=8.1.0",
    "pydantic>=2.10.5",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.6",
    "mypy>=1.14.1",
]

[project.scripts]
gworkspace-mcp = "gworkspace_mcp.cli.main:main"
workspace = "gworkspace_mcp.cli.main:main"

[project.urls]
Homepage = "https://github.com/masapasa/google-workspace-mcp"
Documentation = "https://github.com/masapasa/google-workspace-mcp#readme"
Repository = "https://github.com/masapasa/google-workspace-mcp"
Issues = "https://github.com/masapasa/google-workspace-mcp/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (may require mocked services)",
    "slow: Slow tests that should be run sparingly",
]
filterwarnings = [
    "ignore::DeprecationWarning:google.*:",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.3"
version_files = [
    "VERSION",
    "src/gworkspace_mcp/VERSION",
    "src/gworkspace_mcp/__version__.py:__version__"
]
tag_format = "v$version"
update_changelog_on_bump = true
changelog_incremental = true

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/VERSION",
    "/BUILD_NUMBER",
]

[tool.hatch.build.targets.wheel]
packages = ["src/gworkspace_mcp"]
